Move session invalidation to the child (session) route
